requirements.txt
pandas==1.4.2
google-cloud-speech==2.15.1